+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- UnrecognizedVersionError#version returns the version number that wasn't recognized
+- UnrecognizedVersionError#version returns the version number that wasn't recognized.
+- UnrecognizedFactionError, which is raised by CardAndCount#new (and by extension, CardSet#from_deck_code) if an unrecognized faction number is encountered.
 
 ## [0.1.0] - 2020-08-28
 ### Added